+++ GUI/frames/config_frame/sim_config_frame/gui_sim_config_frame.h
@@ -75,8 +75,8 @@ namespace RhodotronSimulatorGUI::frames::subframes{
         std::string ProduceSimConfiguration();
         void SetSimConfiguration(std::string config);
 
-        const double GetStartTime(){ return (_t0_entry ? _t0_entry->GetNumber() : 0);}
-        const double GetEndTime(){ return (_tf_entry ? _tf_entry->GetNumber() : 0);}
+        const float GetStartTime(){ return (_t0_entry ? _t0_entry->GetNumber() : 0);}
+        const float GetEndTime(){ return (_tf_entry ? _tf_entry->GetNumber() : 0);}
     };
 }
